bindings/qplayer: rename subfolder to torch
